add `frames` parameter to chunks(); ...
- fix up some of the class documentation strings

- upgrade `frames` and `frame_count` to `long long` so frames is not
  limited to 3hrs @ 192 kHz

+++ src/py_pastream.c
@@ -5,7 +5,7 @@
 
 void init_stream(
     Py_PaBufferedStream *stream, 
-    long frames,
+    long long frames,
     long pad,
     unsigned long offset,
     PaUtilRingBuffer *rxbuff,
@@ -48,7 +48,7 @@ int callback(
     unsigned long frames_left = frame_count, offset = 0;
     ring_buffer_size_t oframes, iframes;
     Py_PaBufferedStream *stream = (Py_PaBufferedStream *) user_data;
-    long frames = stream->frames;
+    long long frames = stream->frames;
     long pad = stream->pad;
 
     if ( status & 0xF ) {
