+++ src/py_pastream.h
@@ -6,13 +6,13 @@ typedef struct Py_PaBufferedStream {
     PaStreamCallbackTimeInfo* lastTime;
     int last_callback;
     unsigned char __autoframes;   // Internal use only
-    long frames;                  // Number of frames to play/record (0 means infinite)
+    long long frames;             // Number of frames to play/record (-1 means infinite)
     long pad;                     // Number of zero frames to pad the playback with
                                   // (< 0 means to pad playback whenever buffer is empty)
     unsigned long xruns;
     unsigned long inputOverflows, inputUnderflows;
     unsigned long outputOverflows, outputUnderflows;
-    unsigned long frame_count;    // Number of frames successfully processed
+    unsigned long long frame_count;        // Number of frames successfully processed
     unsigned long offset;         // Number of frames to skip from beginning of recordings
     PaUtilRingBuffer* rxbuff;     // Receive buffer
     PaUtilRingBuffer* txbuff;     // Transmit buffer
@@ -22,7 +22,7 @@ typedef struct Py_PaBufferedStream {
 // call once for initialization
 void init_stream(
     Py_PaBufferedStream *stream, 
-    long frames,
+    long long frames,
     long pad,
     unsigned long offset,
     PaUtilRingBuffer *rxbuff,
